requirements.txt
flask==3.0.0
flask-socketio==5.3.5
requests==2.31.0
websocket-client==1.7.0
python-socketio==5.10.0
python-dotenv==1.0.0
openai>=1.0.0
